refactor: remove library
